app/voice-ai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-ai"
version = "1.0.0"
requires-python = ">=3.10, <3.15"

dependencies = [
    "kittentts",
    "livekit-agents[openai,silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "mlx-audio>=0.2.10",
    "numpy>=2.2.6",
    "ollama>=0.6.1",
    "onnxruntime>=1.24.3",
    "python-dotenv",
    "requests>=2.32.5",
    "setuptools>=82.0.1",
    "sounddevice>=0.5.5",
    "webrtcvad-wheels>=2.0.14",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv.sources]
kittentts = { url = "https://github.com/KittenML/KittenTTS/releases/download/0.8.1/kittentts-0.8.1-py3-none-any.whl" }
